requirements.txt
cryptography
pycryptodome
ecdsa